sql: move struct port_sql definition to sql/port.h
Needed for #6896

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring
